change(architecture): 领域类型枚举、ID 转换、哨兵错误、接口拆分
变更点：
- 为 task/identity/pomodoro 域添加类型安全枚举（TaskState/TaskPriority/UserRole/PomodoroType/ViewType），替换魔法 uint8/string
- 引入 idutil.ParseID/FormatID 消除 application 层 70+ 处 strconv 样板
- 创建 domain/errors 哨兵错误（ErrInvalidUserID 等 14 个），application 层添加 %w 链式包装
- 拆分 TaskApp 接口为 TaskApp/TaskCheckItemApp/TaskCommentApp，控制器按需依赖
- 安装 golangci-lint v1.64.8，修复 .golangci.yml 配置与 gofmt 格式

其他:
- ViewType 作为共享类型定义在 domain/types，供 project 和 tag 偏好共用
- persistence 层添加 uint8()/entities.TaskState() 等显式类型转换

+++ application/user/converters.go
@@ -16,7 +16,7 @@ func UserEntity2Res(e *entities.User) *types.GetUserProfileRes {
 	res.Avatar = e.Avatar
 	res.Role = ""
 	res.CreatedFrom = e.CreatedFrom
-	res.State = e.State
+	res.State = uint8(e.State)
 	res.Config = e.Config
 	res.CreatedAt = e.CreatedAt.Format(time.RFC3339)
 	res.UpdatedAt = e.UpdatedAt.Format(time.RFC3339)
